Fix a bug in retrieving the correct spam filter class
Perhaps there was something in an earlier version of the code before I
started refactoring it where it worked, or something about the refactor
that caused it to stop working.

Regardless, after increasing the test data used to test the spam filter
classes, it became clear that the default, Null spam filter was being
returned in most, if not all cases.

So, this change corrects that so that the correct spam filter class is
returned wherever it is required and the null class is only a last
resort backup option.